Updated Shape.java and Star.java

+++ Team33CMPUT301F21T33Wednesday/app/src/main/java/com/example/team33cmput301f21t33wednesday/Star.java
@@ -5,7 +5,7 @@ public class Star extends Shape {
     private int pointAngle;
     private String colour;
 
-    public Star(int numPoints, float pointAngle, String colour) {
+    public Star(int numPoints, int pointAngle, String colour) {
         this.numPoints = numPoints;
         this.pointAngle = pointAngle;
         this.colour = colour;
